requirements.txt
binance-futures-connector==4.0.0
celery==5.4.0
Django==5.0.4
django-celery-beat==2.6.0
django-redis==5.4.0
flower==2.0.1
gunicorn==22.0.0
ipython==8.24.0
psycopg2-binary==2.9.9
websocket-client==1.8.0
djangorestframework==3.15.2
django-filter==24.3
Markdown==3.6
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
